Frontend refinement

See merge request misit-bachelor/pacs2go!14

+++ pacs2go/frontend/pages/project.py
@@ -1,9 +1,21 @@
-from dash import html, callback, Input, Output, register_page, ctx, State, no_update
-from pacs2go.data_interface.pacs_data_interface import Project
-import dash_bootstrap_components as dbc
-from pacs2go.frontend.helpers import get_connection, colors
+from typing import Optional
 
-register_page(__name__, title='Project',
+import dash_bootstrap_components as dbc
+from dash import callback
+from dash import ctx
+from dash import dcc
+from dash import html
+from dash import Input
+from dash import no_update
+from dash import Output
+from dash import register_page
+from dash import State
+
+from pacs2go.data_interface.pacs_data_interface import Project
+from pacs2go.frontend.helpers import colors
+from pacs2go.frontend.helpers import get_connection
+
+register_page(__name__, title='Project - PACS2go',
               path_template='/project/<project_name>')
 
 
@@ -14,11 +26,11 @@ def get_details(project: Project):
 
 
 def get_directories(project: Project):
-    # get list of all directory names and number of files per directory
+    # Get list of all directory names and number of files per directory
     rows = []
     for d in project.get_all_directories():
-        # directory names represent links to individual directory pages
-        rows.append(html.Tr([html.Td(html.A(d.name, href="", className="text-decoration-none", style={'color': colors['links']})), html.Td(
+        # Directory names represent links to individual directory pages
+        rows.append(html.Tr([html.Td(dcc.Link(d.name, href=f"/dir/{project.name}/{d.name}", className="text-decoration-none", style={'color': colors['links']})), html.Td(
             len(d.get_all_files()))]))
 
     table_header = [
@@ -28,19 +40,19 @@ def get_directories(project: Project):
 
     table_body = [html.Tbody(rows)]
 
-    # put together directory table
+    # Put together directory table
     table = dbc.Table(table_header + table_body,
                       striped=True, bordered=True, hover=True)
     return html.Div([html.H5("Directories:"), table])
 
 
 def modal_delete(project: Project):
-    # modal view for project deletion
+    # Modal view for project deletion
     return html.Div([
-        # button which triggers modal activation
+        # Button which triggers modal activation
         dbc.Button([html.I(className="bi bi-trash me-2"),
                     "Delete Project"], id="delete_project", size="md", color="danger"),
-        # actual modal view
+        # Actual modal view
         dbc.Modal(
             [
                 dbc.ModalHeader(dbc.ModalTitle(
@@ -52,10 +64,10 @@ def modal_delete(project: Project):
                     dbc.Input(id="project", value=project.name, disabled=True),
                 ]),
                 dbc.ModalFooter([
-                    # button which triggers the deletion of a project (see modal_and_project_creation)
+                    # Button which triggers the deletion of a project (see modal_and_project_creation)
                     dbc.Button("Delete Project",
                                id="delete_and_close", color="danger"),
-                    # button which causes modal to close/disappear
+                    # Button which causes modal to close/disappear
                     dbc.Button("Close", id="close_modal_delete"),
                 ]),
             ],
@@ -66,12 +78,12 @@ def modal_delete(project: Project):
 
 
 def modal_delete_data(project: Project):
-    # modal view for deleting all directories of a project
+    # Modal view for deleting all directories of a project
     return html.Div([
-        # button which triggers modal activation
+        # Button which triggers modal activation
         dbc.Button([html.I(className="bi bi-trash me-2"),
                     "Delete All Directories"], id="delete_project_data", size="md", color="danger"),
-        # actual modal view
+        # Actual modal view
         dbc.Modal(
             [
                 dbc.ModalHeader(dbc.ModalTitle(
@@ -84,10 +96,10 @@ def modal_delete_data(project: Project):
                               value=project.name, disabled=True),
                 ]),
                 dbc.ModalFooter([
-                    # button which triggers the directory deletion (see modal_and_project_creation)
+                    # Button which triggers the directory deletion (see modal_and_project_creation)
                     dbc.Button("Delete All Directories",
                                id="delete_data_and_close", color="danger"),
-                    # button which causes modal to close/disappear
+                    # Button which causes modal to close/disappear
                     dbc.Button("Close", id="close_modal_delete_data"),
                 ]),
             ],
@@ -98,6 +110,7 @@ def modal_delete_data(project: Project):
 
 
 def insert_data(project: Project):
+    # Link to Upload functionality with a set project name
     return html.Div(dbc.Button([html.I(className="bi bi-cloud-upload me-2"),
                     "Insert Data"], href=f"/upload/{project.name}", size="md", color="success"))
 
@@ -106,50 +119,60 @@ def insert_data(project: Project):
 #   Callbacks   #
 #################
 
-# callback for project deletion modal view and executing project deletion
 @callback([Output('modal_delete', 'is_open'), Output('delete-project-content', 'children')],
           [Input('delete_project', 'n_clicks'), Input(
               'close_modal_delete', 'n_clicks'), Input('delete_and_close', 'n_clicks')],
           State("modal_delete", "is_open"), State("project", "value"))
+# Callback for project deletion modal view and executing project deletion
 def modal_and_project_deletion(open, close, delete_and_close, is_open, project_name):
-    # open/close modal via button click
+    # Open/close modal via button click
     if ctx.triggered_id == "delete_project" or ctx.triggered_id == "close_modal_delete":
         return not is_open, no_update
+
     if ctx.triggered_id == "delete_and_close":
         try:
             with get_connection() as connection:
-                p = connection.get_project(project_name)
-                p.delete_project()
-                # TODO: redirect to project list view
-                return not is_open, no_update
+                project = connection.get_project(project_name)
+
+                if project:
+                    project.delete_project()
+
+                # Redirect to project list after deletion
+                return not is_open, dcc.Location(href=f"/projects/", id="redirect_after_project_delete")
+
         except Exception as err:
             return is_open, dbc.Alert("Can't be deleted " + str(err), color="danger")
     else:
         return is_open, no_update
 
 
-# callback used to delete all directories of a project (open modal view + actual deletion)
 @callback([Output('modal_delete_data', 'is_open'), Output('delete-project-data-content', 'children')],
           [Input('delete_project_data', 'n_clicks'), Input(
               'close_modal_delete_data', 'n_clicks'), Input('delete_data_and_close', 'n_clicks')],
           State("modal_delete_data", "is_open"), State("project_2", "value"))
+# Callback used to delete all directories of a project (open modal view + actual deletion)
 def modal_and_project_data_deletion(open, close, delete_data_and_close, is_open, project_name):
-    # open/close modal via button click
+    # Open/close modal via button click
     if ctx.triggered_id == "delete_project_data" or ctx.triggered_id == "close_modal_delete_data":
         return not is_open, no_update
+
+    # Delete Button in Modal View
     if ctx.triggered_id == "delete_data_and_close":
         try:
             with get_connection() as connection:
-                p = connection.get_project(project_name)
-                dirs = p.get_all_directories()
-                if len(dirs) == 0:
-                    return is_open,  dbc.Alert("Project is empty", color="danger")
-                else:
-                    for d in dirs:
-                        d.delete_directory()
-                    return not is_open, no_update
+                project = connection.get_project(project_name)
+                if project:
+                    dirs = project.get_all_directories()
+                    if len(dirs) == 0:
+                        return is_open,  dbc.Alert("Project is empty", color="danger")
+                    else:
+                        for d in dirs:
+                            d.delete_directory()
+                        return not is_open, no_update
+
         except Exception as err:
             return is_open, dbc.Alert("Can't be deleted " + str(err), color="danger")
+
     else:
         return is_open, no_update
 
@@ -158,21 +181,31 @@ def modal_and_project_data_deletion(open, close, delete_data_and_close, is_open,
 #  Page Layout  #
 #################
 
-def layout(project_name=None):
+def layout(project_name: Optional[str] = None):
     try:
-        with get_connection() as connection:
-            project = connection.get_project(project_name)
+        if project_name:
+            with get_connection() as connection:
+                project = connection.get_project(project_name)
+
+                if project:
+                    return html.Div([
+                        # Header including page title and action buttons
+                        dbc.Row([
+                            dbc.Col(html.H1(f"Project {project.name}", style={
+                                    'textAlign': 'left', })),
+                            dbc.Col(
+                                [insert_data(project),
+                                 modal_delete(project),
+                                 modal_delete_data(project), ], className="d-grid gap-2 d-md-flex justify-content-md-end"),
+                        ], className="mb-3"),
+                        # Project Information (owners,..)
+                        get_details(project),
+                        # Directories Table
+                        get_directories(project)
+                    ])
+
+                else:
+                    return dbc.Alert("No Project found.", color="danger")
+
     except:
-        return dbc.Alert("No Project found", color="danger")
-    return html.Div([
-        dbc.Row([
-            dbc.Col(html.H1(f"Project {project.name}", style={
-                    'textAlign': 'left', })),
-            dbc.Col(
-                [insert_data(project),
-                 modal_delete(project),
-                 modal_delete_data(project), ], className="d-grid gap-2 d-md-flex justify-content-md-end"),
-        ], className="mb-3"),
-        get_details(project),
-        get_directories(project)
-    ])
+        return dbc.Alert("No Project found.", color="danger")
